+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Change ZeroTier VPN to not be enabled by default.
 - Add game-devices-udev for supporting more controllers.
 - Change PolyMC to Prism Launcher for Minecraft.
+- Add the 'wl' Broadcom Wi-Fi driver.
 
 ## 3.1.1 - The Boot Fix Update